+++ SmartCityExploration/SmartCityExplorationTests/Feature/ViewModel/CitiesViewModelTests.swift
@@ -206,57 +206,3 @@ final class CitiesViewModelTests: XCTestCase {
     }
 }
 
-final class MockCoreDataStore: CitiesPersistentStore {
-    var retrievalResult: RetrievalResult!
-    var insertionResult: InsertionResult!
-    var favoritesResult: Bool!
-
-    func retrieve() async -> RetrievalResult {
-        guard let retrievalResult else {
-            fatalError("Result was not set, developer error")
-        }
-        return retrievalResult
-    }
-    
-    func insert(items: [SmartCityExploration.CityItem], timestamp: Date) async -> InsertionResult {
-        guard let insertionResult else {
-            fatalError("Result was not set, developer error")
-        }
-        return insertionResult
-    }
-    
-    func addAsFavorite(_ item: SmartCityExploration.CityItem) async -> Bool {
-        guard let favoritesResult else {
-            fatalError("Result was not set, developer error")
-        }
-        return favoritesResult
-    }
-    
-    func removeAsFavorite(_ item: SmartCityExploration.CityItem) async -> Bool {
-        guard let favoritesResult else {
-            fatalError("Result was not set, developer error")
-        }
-        return favoritesResult
-    }
-}
-
-final class MockCitiesAPIClient: CitiesAPIClientProtocol {
-    var result: CitiesAPIClientProtocol.Result!
-    func retrieveCities() async -> CitiesAPIClientProtocol.Result {
-        guard let result else {
-            fatalError("Result was not set, develoepr error")
-        }
-        return result
-    }
-}
-
-final class MockSearchHelper: SearchHelperProtocol {
-    func initiate(with items: [SmartCityExploration.CityItem]) {
-        //
-    }
-    
-    func search(text: String) {
-        //
-    }
-
-}
