+++ SmartCityExploration/SmartCityExplorationTests/Feature/ViewModel/Helpers/MockCitiesAPIClient.swift
@@ -1,0 +1,18 @@
+//
+//  MockCitiesAPIClient.swift
+//  SmartCityExploration
+//
+//  Created by Luis Segovia on 24/03/25.
+//
+
+@testable import SmartCityExploration
+
+final class MockCitiesAPIClient: CitiesAPIClientProtocol {
+    var result: CitiesAPIClientProtocol.Result!
+    func retrieveCities() async -> CitiesAPIClientProtocol.Result {
+        guard let result else {
+            fatalError("Result was not set, develoepr error")
+        }
+        return result
+    }
+}
