+++ SmartCityExploration/SmartCityExplorationTests/Feature/ViewModel/Helpers/MockSearchHelper.swift
@@ -1,0 +1,19 @@
+//
+//  MockSearchHelper.swift
+//  SmartCityExploration
+//
+//  Created by Luis Segovia on 24/03/25.
+//
+
+@testable import SmartCityExploration
+
+final class MockSearchHelper: SearchHelperProtocol {
+    func initiate(with items: [SmartCityExploration.CityItem]) {
+        //
+    }
+    
+    func search(text: String) {
+        //
+    }
+
+}
